// WeatherApp/src/shared/model/types/weatherTypes.ts
export interface IWeather {
    location: Location
    current: Current
    forecast: Forecast | null
    coords: Coords | string
}

export interface Location {
    id?: number
    name: string
    region: string
    country: string
    lat: number
    lon: number
    tz_id?: string
    localtime_epoch?: number
    localtime?: string
    url?: string
}
export interface Coords {
    lat: number, lon: number
}

export interface Current {
    last_updated_epoch: number
    last_updated: string
    temp_c: number
    temp_f: number
    is_day: number
    condition: Condition
    wind_mph: number
    wind_kph: number
    wind_degree: number
    wind_dir: string
    pressure_mb: number
    pressure_in: number
    precip_mm: number
    precip_in: number
    humidity: number
    cloud: number
    feelslike_c: number
    feelslike_f: number
    windchill_c: number
    windchill_f: number
    heatindex_c: number
    heatindex_f: number
    dewpoint_c: number
    dewpoint_f: number
    vis_km: number
    vis_miles: number
    uv: number
    gust_mph: number
    gust_kph: number
}

export interface Condition {
    text: string
    icon: string
    code: number
}

export interface Forecast {
    forecastday: Forecastday[]
}

export interface Forecastday {
    date: string
    date_epoch: number
    day: Day
    astro: Astro
    hour: Hour[]
}